+++ app/src/main/java/fi/tv7/taivastv7/adapter/CategoryGridAdapter.java
@@ -21,6 +21,7 @@ import fi.tv7.taivastv7.R;
 import fi.tv7.taivastv7.helpers.Utils;
 
 import static fi.tv7.taivastv7.helpers.Constants.CAPTION;
+import static fi.tv7.taivastv7.helpers.Constants.EMPTY;
 import static fi.tv7.taivastv7.helpers.Constants.IMAGE_PATH;
 import static fi.tv7.taivastv7.helpers.Constants.LOG_TAG;
 import static fi.tv7.taivastv7.helpers.Constants.ONE_STR;
@@ -134,8 +135,11 @@ public class CategoryGridAdapter extends RecyclerView.Adapter<CategoryGridAdapte
                 }
 
                 value = Utils.getValue(obj, CAPTION);
-                if (value != null) {
+                if (value != null && value.length() > 0) {
                     holder.caption.setText(value);
+                }
+                else {
+                    holder.caption.setText(EMPTY);
                 }
             }
         }