+++ app/src/main/java/fi/tv7/taivastv7/adapter/GuideGridAdapter.java
@@ -22,6 +22,7 @@ import fi.tv7.taivastv7.R;
 import fi.tv7.taivastv7.helpers.Utils;
 
 import static fi.tv7.taivastv7.helpers.Constants.CAPTION;
+import static fi.tv7.taivastv7.helpers.Constants.EMPTY;
 import static fi.tv7.taivastv7.helpers.Constants.IMAGE_PATH;
 import static fi.tv7.taivastv7.helpers.Constants.ONE_STR;
 import static fi.tv7.taivastv7.helpers.Constants.ONGOING_PROGRAM;
@@ -148,8 +149,11 @@ public class GuideGridAdapter extends RecyclerView.Adapter<GuideGridAdapter.Simp
                 }
 
                 value = Utils.getValue(obj, CAPTION);
-                if (value != null) {
+                if (value != null && value.length() > 0) {
                     holder.caption.setText(value);
+                }
+                else {
+                    holder.caption.setText(EMPTY);
                 }
             }
         }